Refactor TimeBox and TeamTimer out of index.ios.js
